sidebar: update rss href
Previously the RSS `href` would evaluate to the following:

If `rss_section` is set to `foo`, the RSS `href` is `/foo/index.xml`.
If `rss_section` is unset, the RSS `href` is `//index.xml`.

This patch moves the `/` into the conditional so that the RSS `href` now
evalates to the following:

If `rss_section` is set to `foo`, the RSS `href` is `/foo/index.xml`.
If `rss_section` is unset, the RSS `href` is `/index.xml`.